// src/modules/merchandising/reports/work-order/yarn-twisting-wrok-order-rpt/yarn-twisting-wrok-order-rpt-type.ts
export type YarnTwistingWorkOrderReportType = {
  ID: number;
  GROUP_COMPANY_NAME: string;
  GROUP_COMPANY_ADDRESS: string;
  SUB_YARN_ID: string;
  SUB_YARN: string;
  SUB_YARN_QUANTITY: string;
  WORK_ORDER_NO: string;
  COMPANY_ID: string;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COMPANY_EMAIL: string;
  COMPANY_WEB: string;
  COMPANY_REMARKS: string;
  SUPPLIER_ID: string;
  SUPPLIER_NAME: string;
  SUPPLIER_ADDRESS: string;
  ISSUE_DATE: string;
  IS_APPROVE: string;
  APPROVE_DATE: string;
  REMARKS: string;
  RECEIVE_STORE_ID: string;
  STORE_PERSONNEL: string;
  MOBILE_NO: string;
  RCV_STORE: string;
  DELIVERY_DATE: string;
  WO_SUBJECT: string;
  ATTENTION: string;
  SUPPLIER_MOBILE_NO: string;
  TOTAL_AMOUNT: number;
  BUYER_ID: string;
  PO_ID: string;
  BUYER_NAME: string;
  PO_NO: string;
  STYLE_ID: string;
  STYLENAME: string;
  MATERIAL_ID: string;
  MTL_NAME: string;
  GMT_COLOR_ID: string;
  GMT_SIZE_ID: string;
  MTL_COLOR_ID: string;
  UOM_ID: string;
  UOM: string;
  SUPPLIER_RATE_PER_PCS: string;
  TOTAL_SUP_AMOUNT: string;
  WORK_ORDER_QTY: string;
  GMT_COLOR_NAME: string;
  GMT_COLOR_CODE: string;
  MTL_COLOR_NAME: string;
  GMT_SIZE_NAME: string;
  DESCRIPTION: string;
  SORTINGNO: string;
  FACTORY_CODE: string;
  NOTE: string;
  APPROVED_BY: string;
  APPROVED_BY_DESG: string;
  PREPARED_BY: string;
  PREPARED_BY_DESG: string;
  PREPARED_BY_MOBILENO: string;
};